+++ inc/tonote/index.ts
@@ -1,6 +1,8 @@
 
 import Token from "./token";
 
+export * from "./blockizer";
+
 
 
 export {